Squashed 'charset/' changes from 443a213..0bcab4b
0bcab4b gcc warning fix: add missing break in read_iso2022.
10975d5 gcc warning fixes relating to integer types.
7858df7 gcc warning fix: mark deliberate case fallthroughs.
cadc277 gcc warning fix: cast function pointers via void (*)(void).

git-subtree-dir: charset
git-subtree-split: 0bcab4bd82e99601609ba48287c5f8418f9cc56b

+++ confuse.c
@@ -55,7 +55,7 @@ int main(int argc, char **argv)
             else
                 base = 10;
             semi_ok = true;
-        } else if (mbtowc(&ch, p, strlen(p)) == strlen(p)) {
+        } else if ((size_t)mbtowc(&ch, p, strlen(p)) == strlen(p)) {
             chars[nchars++] = ch;
             continue;
         }
